split discord command handling into separate file
