+++ src/bot.ts
@@ -36,6 +36,7 @@ import { Log } from "./log";
 import * as Discord from "discord.js";
 import * as mime from "mime";
 import { IMatrixEvent, IMatrixMediaInfo } from "./matrixtypes";
+import { DiscordCommandHandler } from "./discordcommandhandler";
 
 const log = new Log("DiscordBot");
 
@@ -74,6 +75,7 @@ export class DiscordBot {
     private channelSync: ChannelSyncroniser;
     private roomHandler: MatrixRoomHandler;
     private provisioner: Provisioner;
+    private discordCommandHandler: DiscordCommandHandler;
     /* Caches */
     private roomIdsForGuildCache: Map<string, {roomIds: string[], ts: number}> = new Map();
 
@@ -99,6 +101,7 @@ export class DiscordBot {
             new MatrixEventProcessorOpts(config, bridge, this),
         );
         this.channelSync = new ChannelSyncroniser(bridge, config, this, store.roomStore);
+        this.discordCommandHandler = new DiscordCommandHandler(this);
         // init vars
         this.sentMessages = [];
         this.discordMessageQueue = {};
@@ -711,7 +714,7 @@ export class DiscordBot {
 
         // check if it is a command to process by the bot itself
         if (msg.content.startsWith("!matrix")) {
-            await this.roomHandler.HandleDiscordCommand(msg);
+            await this.discordCommandHandler.Process(msg);
             return;
         }
